src/requirements.txt
Django>=4.2,<5.1
wagtail>=6.2,<6.3
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.3.1
django-filter>=24.3